requirements.txt
contractions
streamlit
pandas
scikit-learn
nltk
psycopg2-binary
